requirements.txt
chromadb==0.5.4
pypdf==4.3.1
pdfplumber==0.11.4
pdfminer.six==20231228
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
rank-bm25==0.2.2
sentence-transformers==3.0.1
nltk==3.9.1
ollama==0.3.3
rapidfuzz==3.9.6
regex==2024.7.24
plotly==5.23.0
streamlit==1.37.1
pillow==10.4.0
pytesseract==0.3.13
# 👇 removed camelot-py[cv] to avoid pdftopng issue on Win+Py3.12
# camelot-py[cv]==0.11.0
